+++ src/main/java/com/se07/controller/controllers/ControllerTaoBieuMauDipDacBietCanBoView.java
@@ -1,9 +1,6 @@
 package com.se07.controller.controllers;
 
 import javafx.event.ActionEvent;
-import javafx.scene.Node;
-
-import java.io.IOError;
 import java.io.IOException;
 
 public class ControllerTaoBieuMauDipDacBietCanBoView extends ControllerTrangChuView{
