--- src/main/java/com/se07/controller/controllers/ControllerTrangChu.java
@@ -1,4 +1,0 @@
-package com.se07.controller.controllers;
-
-public class ControllerTrangChu {
-}
