+++ src/main/java/com/se07/controller/controllers/ControllerTrangChuView.java
@@ -1,23 +1,15 @@
 package com.se07.controller.controllers;
 import com.se07.view.*;
-import eu.hansolo.tilesfx.Alarm;
-import javafx.application.Platform;
 import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
-import javafx.stage.StageStyle;
-
 import java.io.IOException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
+
 import java.util.ResourceBundle;
 
 public class ControllerTrangChuView implements Initializable{
