--- src/main/java/com/se07/view/AdminView.java
@@ -1,11 +1,0 @@
-package com.se07.view;
-
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.image.Image;
-import javafx.stage.Stage;
-import javafx.stage.StageStyle;
-
-public class AdminView {
-
-}
