+++ src/main/java/com/se07/view/TrangChuCanBoView.java
@@ -9,5 +9,22 @@ import javafx.stage.Stage;
 import javafx.stage.StageStyle;
 
 public class TrangChuCanBoView {
+    public Stage stage;
+    public Stage openWindow() {
+        try{//
+            stage = new Stage();
+            FXMLLoader fxmlLoader = new FXMLLoader(LoginView.class.getResource("TrangChuCanBoView.fxml"));
+            Scene scene = new Scene(fxmlLoader.load(), 915, 603);
+            stage.setTitle("QUẢN LÝ NHÂN KHẨU");
+            stage.initStyle(StageStyle.UNDECORATED);
+            stage.setScene(scene);
+            stage.show();
+        }
+        catch (Exception e){
+            System.out.println(1);
+            System.out.println(e.getMessage());
+        }
+        return null;
+    }
 
 }
