+++ src/main/resources/com/se07/view/Css/treasurer.css
@@ -1,0 +1,22 @@
+#buttonHomeTreasurer{
+    -fx-background-color : rgb(40,132,217);
+}
+#buttonHomeTreasurer:hover{
+    -fx-background-color: rgba(0,0,0);
+}
+#buttonPeopleTreasurer{
+    -fx-background-color: rgba(0, 53, 240, 0.7);
+}
+#buttonPeopleTreasurer:hover{
+    -fx-background-color:rgba(4, 57, 244, 0.95)
+}
+
+#buttonAwardTreasurer{
+    -fx-background-color: black;
+}
+#buttonAwardTreasurer:hover{
+    -fx-background-color: #177306;
+}
+#paneRightTreasurer{
+    -fx-background-color: linear-gradient(#7758D1 2%,  #F7CBFD 40%);
+}
