+++ src/main/resources/com/se07/view/Css/user.css
@@ -1,0 +1,102 @@
+#buttonHomeUser{
+    -fx-background-color : rgb(40,132,217);
+}
+#buttonHomeUser:hover{
+    -fx-background-color: rgba(0,0,0);
+}
+#buttonPeopleUser{
+    -fx-background-color: rgba(0, 53, 240, 0.7);
+}
+#buttonPeopleUser:hover{
+    -fx-background-color:rgba(4, 57, 244, 0.95)
+}
+
+#buttonAwardUser{
+    -fx-background-color: rgba(0, 53, 240, 0.7);
+}
+#buttonAwardUser:hover{
+    -fx-background-color:rgba(4, 57, 244, 0.95);
+}
+#paneRightUser{
+    -fx-background-color: linear-gradient(#7758D1 2%,  #F7CBFD 40%);
+}
+#menuButtonAwardUser{
+    -fx-background-color: rgba(0, 53, 240, 0.7);
+}
+#menuButtonAwardUser:hover{
+    -fx-background-color: rgba(4, 57, 244, 0.95);
+}
+#paneNewPeopleUser{
+    -fx-background-color: linear-gradient(#7758D1 2%,  #F7CBFD 40%);
+}
+#buttonCancelTemporaryRegistrationPeopleUser{
+    -fx-background-color: black;
+}
+#buttonCancelTemporaryRegistrationPeopleUser:hover{
+    -fx-background-color: #910c0c;
+}
+#buttonAcceptTemporaryRegistrationPeopleUser{
+    -fx-background-color: black;
+}
+#buttonAcceptTemporaryRegistrationPeopleUser:hover{
+    -fx-background-color: #177306;
+
+}
+
+#buttonNewPeopleUser{
+    -fx-background-color: black;
+}
+#buttonNewPeopleUser:hover{
+    -fx-background-color: #BF3FBF;
+}
+#buttonLogoutUser{
+    -fx-background-color: black;
+}
+#buttonLogoutUser:hover{
+    -fx-background-color: #BF3FBF;
+}
+#buttonSplitUserUser{
+    -fx-background-color: black;
+}
+#buttonSplitUserUser:hover{
+    -fx-background-color: #BF3FBF;
+}
+#buttonRemoveUserUser{
+    -fx-background-color: black;
+}
+#buttonRemoveUserUser:hover{
+    -fx-background-color: #BF3FBF;
+}
+#paneSplitUserUser{
+    -fx-background-color: linear-gradient(#7758D1 2%,  #F7CBFD 40%);
+}
+#anchorPaneImportAwardUser{
+    -fx-background-color: linear-gradient(#7758D1 2%,  #F7CBFD 40%);
+}
+#anchorPaneCheckAwardUser{
+    -fx-background-color: linear-gradient(#7758D1 2%,  #F7CBFD 40%);
+}
+#buttonCreateImportFormAwardUser{
+    -fx-background-color: black;
+}
+#buttonCreateImportFormAwardUser:hover{
+    -fx-background-color: #177306;
+}
+#buttonCancelImportFormAwardUser{
+    -fx-background-color: black;
+}
+#buttonCancelImportFormAwardUser:hover{
+    -fx-background-color: #910c0c;
+}
+#buttonAcceptSplitUserUser{
+    -fx-background-color: black;
+}
+#buttonAcceptSplitUserUser:hover{
+    -fx-background-color: #177306;
+}
+#buttonCancelSplitUserUser{
+    -fx-background-color: black;
+}
+#buttonCancelSplitUserUser:hover {
+    -fx-background-color: #910c0c;
+}
